Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# PDF
PyMuPDF==1.24.5

# Vector store
chromadb==0.5.3

# HTTP
requests==2.32.3

# Settings (optional but handy)
pydantic-settings==2.3.4

# Database
SQLAlchemy>=2.0,<3
# Postgres driver (only needed if you use Postgres)
psycopg2-binary>=2.9